+++ mongodb/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2019-06-16
 ### Changed
 - **BREAKING**: Upgrade base agent library to latest.
 
@@ -34,6 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - OpenTracing integration
 
 
-[Unreleased]: https://github.com/replicante-io/agents/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/replicante-io/agents/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/replicante-io/agents/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/replicante-io/agents/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/replicante-io/agents/compare/v0.1.0...v0.2.0